Functionality revert + lots of comments
fully swapped functionality from creating nodes in bulk back to creating them over time. The bulk method didn't work any differently, but I am leaving the portions in for the future. Found a bug with instancing too many 3D node scenes at once and made a second project as a minimal project to reproduce it. The full solution to that problem isn't beneficial to the problems I am having with latency.