Reverted failed custom song/regex implementation
Reverted the base functionality of the game itself to a state that functions properly. Commented out code that tried to load and open an existing file instead of only being able to use the hardcoded song file path. Commented out code that changed scenes to the "song selection" scene because functionality doesn't work.